BactDating: Bayesian inference of ancestral dates on bacterial phylogenetic trees

BactDating is a R package to perform Bayesian inference of ancestral dates on bacterial phylogenetic trees.

The main functions of the package are:

initRoot to initiate the root of an unrooted tree in a way that maximises the correlation between sampling dates and root-to-tip distances.

roottotip to perform a linear regression analysis between sampling dates and root-to-tip distances.

bactdate to perform Bayesian inference of ancestral dates on a phylogenetic tree.

References

Didelot et al (2018). Bayesian inference of ancestral dates on bacterial phylogenetic trees. Nucleic Acids Research, 46:e134.

See also

Author

Xavier Didelot xavier.didelot@gmail.com