Main function of BactDating package to date the nodes of a phylogenetic tree

Main function of BactDating package to date the nodes of a phylogenetic tree

bactdate(
  tree,
  date,
  initMu = NA,
  initAlpha = NA,
  initSigma = NA,
  updateMu = T,
  updateAlpha = T,
  updateSigma = T,
  updateRoot = T,
  nbIts = 10000,
  thin = ceiling(nbIts/1000),
  useCoalPrior = T,
  model = "arc",
  useRec = F,
  minbralen = 0.1,
  showProgress = F
)

Arguments

tree

Tree wih branches measured in unit of substitutions

date

Sampling dates for the leaves of the tree

initMu

Initial rate of substitutions per genome (not per site), or zero to use root-to-tip estimate

initAlpha

Initial coalescent time unit

initSigma

Initial std on per-branch substitution rate (only used in relaxed or mixed-relaxed models)

updateMu

Whether or not to update the substitution rate

updateAlpha

Whether or not to update the coalescent time unit

updateSigma

Whether or not to per-branch substitution rate (only used in relaxed or mixed-relaxed models)

updateRoot

Whether or not to use the root finding algorithm

nbIts

Number of MCMC iterations to perform

thin

Thining interval between recorded MCMC samples

useCoalPrior

Whether or not to use a coalescent prior on the tree

model

Which model to use (poisson or negbin or strictgamma or relaxedgamma or mixedgamma or arc or carc or mixedcarc)

useRec

Whether or not to use results from previous recombination analysis

minbralen

Minimum branch length of the phylogenetic tree (in number of substitutions)

showProgress

Whether or not to show a progress bar

Value

Dating results