Modified drop.tip() from 'ape' package that also updates the recombination proportions added when loaded from CFML or Gubbins

Modified drop.tip() from 'ape' package that also updates the recombination proportions added when loaded from CFML or Gubbins

# S3 method for useRec
drop.tip(
  phy,
  tip,
  trim.internal = TRUE,
  subtree = FALSE,
  root.edge = 0,
  rooted = is.rooted(phy),
  collapse.singles = TRUE,
  interactive = FALSE
)

Arguments

phy

Tree from loadCFML() or loadGubbins()

tip

a vector of mode numeric or character specifying the tips to delete.

trim.internal

a logical specifying whether to delete the corresponding internal branches.

subtree

a logical specifying whether to output in the tree how many tips have been deleted and where.

root.edge

an integer giving the number of internal branches to be used to build the new root edge. This has no effect if trim.internal FALSE.

rooted

a logical indicating whether the tree must be treated as rooted or not. This allows to force the tree to be considered as unrooted.

collapse.singles

a logical specifying whether to delete the internal nodes of degree 2.

interactive

if TRUE the user is asked to select the tips or the node by clicking on the tree which must be plotted.

Value

tree with rec data