Root to tip correlation

Root to tip correlation

roottotip(
  tree,
  date,
  rate = NA,
  permTest = 10000,
  showFig = T,
  colored = T,
  showPredInt = "gamma",
  showText = T,
  showTree = T
)

Arguments

tree

Phylogenetic tree

date

Dates of sampling

rate

Evolutionary rate, estimated unless provided

permTest

Number of permutations to perform to compute the p-value using a permutation test

showFig

Whether or not to show the root-to-tip regression figure

colored

Whether or not to use colors illustrating dates

showPredInt

To show 95percent confidence intervals of a strict clock model, can be 'poisson' or 'gamma'

showText

Whether to show the title and axis labels

showTree

Whether to show the tree or not

Value

List containing estimated clock rate, date of origin and p-value